updated Notes.js and noteRoutes.js

+++ routes/noteRoutes.js
@@ -14,7 +14,7 @@ router.get('/', async (req, res) => {
   // This currently finds all notes in the database.
   // It should only find notes owned by the logged in user.
   try {
-    const notes = await Note.find({});
+    const notes = await Note.find({user: req.user._id});
     res.json(notes);
   } catch (err) {
     res.status(500).json(err);
@@ -27,22 +27,33 @@ router.post('/', async (req, res) => {
     const note = await Note.create({
       ...req.body,
       // The user ID needs to be added here
+      user: req.user._id
     });
     res.status(201).json(note);
   } catch (err) {
     res.status(400).json(err);
   }
 });
+
  
 // PUT /api/notes/:id - Update a note
 router.put('/:id', async (req, res) => {
   try {
     // This needs an authorization check
-    const note = await Note.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const note = await Note.findById(req.params.id);
+
     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
-    res.json(note);
+
+    if(note.user.toString() !== req.user._id){
+      return res.status(403).json({message: 'User is not authorized to update this note'})
+    }
+
+    const updatedNote = await Note.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    
+    res.json(updatedNote);
+
   } catch (err) {
     res.status(500).json(err);
   }
@@ -52,11 +63,35 @@ router.put('/:id', async (req, res) => {
 router.delete('/:id', async (req, res) => {
   try {
     // This needs an authorization check
-    const note = await Note.findByIdAndDelete(req.params.id);
-    if (!note) {
+     const note = await Note.findById(req.params.id);
+
+     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
+
+     if(note.user.toString() !== req.user._id){
+      return res.status(403).json({message: 'User is not authorized to delete this note'})
+    }
+
+    const deletedNote = await Note.findByIdAndDelete(req.params.id);
+    
     res.json({ message: 'Note deleted!' });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+//Get Single Note
+router.get('/:id', async (req, res) => {
+  
+  try {
+
+    const note = await Note.findOne({user: req.user._id, _id:req.params.id});
+
+     if (!note) {
+      return res.status(404).json({ message: "Note not found" });
+    }
+    res.json(note);
   } catch (err) {
     res.status(500).json(err);
   }
